requirements.txt
streamlit
tensorflow-cpu
pandas
numpy
joblib
requests